Rename files: /index.js --> /server.js, /routes/index.js --> api.js, /controllers/index.js --> api.js
Changed:
code error 500 and request format json.

note: included '.DS_Store' in .gitignore.
