+++ app.js
@@ -6,7 +6,7 @@ const express = require('express')
 const bodyParser = require('body-parser')
 
 const app = express()
-const api = require('./routes')
+const api = require('./routes/api')
 
 app.use(bodyParser.urlencoded({extended: false}))
 app.use(bodyParser.json())
