+++ controllers/api.js
@@ -5,36 +5,36 @@ const fs = require('fs')
 function getCollaborators(req, res){
 	fs.readFile('./collaborators.json', 'utf8', (err, data) => {
 		if (err) {
-			return res.status(400).send(err)
+			return res.status(500).json({success: false, message: err.message })
 		}
-		res.status(200).send(data)
+		res.send(data)
 	})
 }
 
 function getCollaborator(req, res){
 	fs.readFile('collaborator_id.json', 'utf8', (err, data) => {
 		if (err) {
-			return res.status(400).send(err)
+			return res.status(500).json({success: false, message: err.message })
 		}
-		res.status(200).send(data)
+		res.send(data)
 	})
 }
 
 function getProjects(req, res){
 	fs.readFile('./projects.json', 'utf8', (err, data) => {
 		if (err) {
-			return res.status(400).send(err)
+			return res.status(500).json({success: false, message: err.message })
 		}
-		res.status(200).send(data)
+		res.send(data)
 	})
 }
 
 function getProject(req, res){
 	fs.readFile('./project_id.json', 'utf8', (err, data) => {
 		if (err) {
-			return res.status(400).send(err)
+			return res.status(500).json({success: false, message: err.message })
 		}
-		res.status(200).send(data)
+		res.send(data)
 	})
 }
